Added Trigger Volumes
PLEASE USE THE METHODS IN PHYSICSENGINE() TO ADD TRIGGER VOLUMES!

TO ADD LOGIC TO THE TRIGGER VOLUME: ColliderCallback.cpp in the onTrigger function are 3 if statements. One for pickup boxes, lap markers and hazards. Please enter program logic there. Ask me if you have any questions!

The sizes of the trigger volumes are kind of weird, you may need to experiment with the dimensions based on the obj file you use!

Sorry, I had to comment out a lot of cout statements for my own debug! The statements are all still there, they are just commented out!